Patch from PR #33443
