--- plugins/catalog-unprocessed-entities/src/alpha/devToolsContent.tsx
@@ -1,38 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { SubPageBlueprint } from '@backstage/frontend-plugin-api';
-import { Content } from '@backstage/core-components';
-
-/**
- * DevTools content for catalog unprocessed entities.
- *
- * @alpha
- */
-export const unprocessedEntitiesDevToolsContent = SubPageBlueprint.make({
-  attachTo: { id: 'page:devtools', input: 'pages' },
-  name: 'unprocessed-entities',
-  params: {
-    path: 'unprocessed-entities',
-    title: 'Unprocessed Entities',
-    loader: () =>
-      import('../components/UnprocessedEntities').then(m => (
-        <Content>
-          <m.UnprocessedEntitiesContent />
-        </Content>
-      )),
-  },
-});
