+++ plugins/catalog-unprocessed-entities/src/alpha/index.ts
@@ -15,4 +15,3 @@
  */
 
 export { default } from './plugin';
-export { unprocessedEntitiesDevToolsContent } from './devToolsContent';
