+++ plugins/catalog-unprocessed-entities/src/alpha/plugin.tsx
@@ -21,6 +21,7 @@ import {
   ApiBlueprint,
   PageBlueprint,
   NavItemBlueprint,
+  SubPageBlueprint,
 } from '@backstage/frontend-plugin-api';
 
 import {
@@ -29,6 +30,7 @@ import {
 } from '../api';
 import QueueIcon from '@material-ui/icons/Queue';
 import { rootRouteRef } from '../routes';
+import { Container } from '@backstage/ui';
 
 /** @alpha */
 export const catalogUnprocessedEntitiesApi = ApiBlueprint.make({
@@ -46,6 +48,7 @@ export const catalogUnprocessedEntitiesApi = ApiBlueprint.make({
 
 /** @alpha */
 export const catalogUnprocessedEntitiesPage = PageBlueprint.make({
+  disabled: true,
   params: {
     path: '/catalog-unprocessed-entities',
     routeRef: rootRouteRef,
@@ -65,6 +68,25 @@ export const catalogUnprocessedEntitiesNavItem = NavItemBlueprint.make({
   },
 });
 
+/**
+ * DevTools content for catalog unprocessed entities.
+ *
+ * @alpha
+ */
+export const unprocessedEntitiesDevToolsContent = SubPageBlueprint.make({
+  attachTo: { id: 'page:devtools', input: 'pages' },
+  params: {
+    path: 'unprocessed-entities',
+    title: 'Unprocessed Entities',
+    loader: () =>
+      import('../components/UnprocessedEntities').then(m => (
+        <Container>
+          <m.UnprocessedEntitiesContent />
+        </Container>
+      )),
+  },
+});
+
 /** @alpha */
 export default createFrontendPlugin({
   pluginId: 'catalog-unprocessed-entities',
@@ -78,5 +100,6 @@ export default createFrontendPlugin({
     catalogUnprocessedEntitiesApi,
     catalogUnprocessedEntitiesPage,
     catalogUnprocessedEntitiesNavItem,
+    unprocessedEntitiesDevToolsContent,
   ],
 });
